Refs #2: move main code to index file
